Remove cookie-setting code
Instead of setting a cookie that bypasses the authorization check, just allow certain paths to be exempt. This is useful for when there's a proxy in the server. You don't want to pass the basic auth information forward to the proxied service, becuase the authorization information may not be the same.